Changed: A wrong Syntax in the Spellhandler caused AE Spells to have no Dropoff. Changed: The Way Viper is calculated within the DoTSpellHandler to fix Viper 1 & 2.

+++ GameServer/spells/SpellHandler.cs
@@ -1677,7 +1677,7 @@ namespace DOL.GS.Spells
 		/// <returns></returns>
 		protected virtual double CalculateAreaVariance(int distance, int radius)
 		{
-			return (double)(distance / radius);
+			return ((double)distance / (double)radius);
 		}
 
 		/// <summary>
